+++ auto_uv/gpu/gpu_vf_curve_applier.py
@@ -333,7 +333,7 @@ def _auto_uv_power_limit_w(runtime_options: dict) -> int | None:
             )
         if value in (None, ""):
             return None
-        power_limit_w = int(round(float(cast(Any, value))))
+        power_limit_w = round(float(cast(Any, value)))
     except (TypeError, ValueError) as exc:
         raise AutoUvError(f"invalid Auto-UV power limit: {value!r}") from exc
     return power_limit_w if power_limit_w > 0 else None
@@ -343,7 +343,7 @@ def _positive_power_limit_w(value: object) -> int | None:
     if value in (None, ""):
         return None
     try:
-        power_limit_w = int(round(float(cast(Any, value))))
+        power_limit_w = round(float(cast(Any, value)))
     except (TypeError, ValueError):
         return None
     return power_limit_w if power_limit_w > 0 else None
